==> 4_vadkarok_hr_parser.ipynb <==
import urllib.request as req
#!pip install fastkml
from fastkml import kml
import pandas as pd
import re

!pip install fastkml

# using the map from https://hargitamegye.ro/vadkarok/interaktiv-terkep.html
# downloading in KML format
url = 'http://www.google.com/maps/d/kml?forcekml=1&mid=1sIXAUcYDAkCBUzw_6DZelr1rt48'

fname, _ = req.urlretrieve(url)
k = kml.KML()

with open(fname, 'rt', encoding='utf-8') as f:
    k.from_string(f.read().encode('utf-8'))

columns_raw = ['geo_loc', 'latitude', 'longitude', 'content', 'image']
df = pd.DataFrame(columns=columns_raw)

folders = next(k.features()).features()

# parse KML, insert data into dataframe
for f in folders:
    for pm in f.features():
        lon, lat, _ =  pm.geometry.coords[0]
        media_ln = ''
        if pm.extended_data is not None:
            media_ln = pm.extended_data.elements[0].value
        df = df.append({
                    'geo_loc': pm.name, 
                    'latitude': lat, 
                    'longitude': lon, 
                    'content': pm.description, 
                    'image': media_ln}, 
                ignore_index=True)

def remove_links(desc):
    '''
        Input: description extracted from KML files
        Output: description without the links of attached images
    '''
    desc = desc.lstrip().rstrip()
    desc = re.sub(r'<img[^>]*/>', '', desc)
    desc = re.sub(r'<br>', '', desc)
    return desc

df['content'] = df['content'].apply(remove_links)

dp = re.compile('([0-9]{4})[. ]+(január|február|március|április|május|június|július|' + 
                'augusztus|szeptember|október|november|december|[0-9]{1,2})[. ]+([0-9]{1,2})')

def standardize_date(date_groups):
    '''
        Input: result of regexp match with 3 groups, corresponding to (year, month, day)
        Output: (year, month, day) tuple, where month is converted to numerical format
    '''
    months_hu = [u'január', u'február', u'március', 
                 u'április', u'május', u'június', 
                 u'július', u'augusztus', u'szeptember', 
                 u'október', u'november', u'december']
    
    y, m, d = date_groups
    
    try: 
        m = str(months_hu.index(m) + 1)
    except:
        return y, m, d
    
    return y, m, d


def get_date(desc):
    '''
        Input: description extracted from KML
        Output: Finds dates in the description, returns it in pandas datetime format
    '''
    res = dp.search(desc)
    if res is not None:
        date = standardize_date(res.groups())
    else:
        date = ''
    
    return pd.to_datetime('-'.join(date))

df['date'] = df['content'].apply(get_date)

lnp = re.compile('https?://[^ ,]*')

def get_link(desc):
    '''
        Input: description extracted from KML
        Output: finds links to external news sites
    '''
    res = lnp.search(desc)
    if res is not None:
        ln = desc[res.start():res.end()]
    else: 
        ln = 'https://hargitamegye.ro/vadkarok/interaktiv-terkep.html'
        
    return ln

df['link'] = df['content'].apply(get_link)

def clean_content(desc):
    '''
        Input: description extracted from KML
        Output: removes links from description
    '''
    if 'http' in desc:
        return desc[:desc.find('http')]
    else:
        return desc

df['content'] = df['content'].apply(clean_content)

print(len(df))

df.head(15)

df['checked']=''
df['exact_loc']=0
df['source']='Vadkárok HR'
df['title']=df['geo_loc']

from utils import text_processor

duplicate=[]
relevant=[]
severity=[]
deaths=[]
for i in df.index:
    l=df.loc[i]['link']
    d=0
    if l:
        if 'szekelyhon' in l: d=1
    duplicate.append(d)
    c=df.loc[i]['content']
    t=df.loc[i]['title']
    r,s,d=text_processor(t,c)
    relevant.append(1) #override r
    severity.append(s)
    deaths.append(d)

df['duplicate']=duplicate
df['relevant']=relevant
df['severity']=severity
df['deaths']=deaths

df=df[['checked',
    'date',
    'link',
    'image',
    'source',
    'title',
    'content',
    'relevant',
    'severity',
    'deaths',
    'duplicate',
    'geo_loc',
    'exact_loc',
    'latitude',
    'longitude']].set_index('checked')

df.to_excel('data/vadkarok_hr.xlsx')
df.to_excel('data/vadkarok_hr_curated.xlsx')

Fill in missing date from image EXIF data

!pip install exifread

no_date=df[df['date'].astype(str)=='NaT']

no_date_yes_image=no_date[no_date['image'].astype(str)!='']

no_date_yes_image

import exifread
import urllib.request

for im_link in no_date_yes_image['image'].values:
    im_link=im_link.split(' ')[0] #if multiple images, take only first
    print(im_link)
    try:
        urllib.request.urlretrieve(im_link, "temp.png")
        with open('temp.png', 'rb') as fh:
            tags = exifread.process_file(fh, stop_tag="EXIF DateTimeOriginal")
            if tags:
                dateTaken = tags["EXIF DateTimeOriginal"]
                print(dateTaken)
            else:
                print('No EXIF')
    except:
        print('No readable image')

Fill in missing dates - if you can, manually
