==> 5_gsheet_pusher.ipynb <==
Open `data/szekelyhon_data.xlsx`,`data/vadkarok_hr.xlsx`, then manually copy and pate into [`GSheet`](https://medve.csaladen.es/data).  
Look at last news items in each category, don't paste any duplicates.  
Be careful to increase cell sizes in _Excel_ for the `date` column, so that it shows the full date. Otherwise it copies only the _Excel_ error `####`.
